endpoint/requirements.txt
flask
werkzeug.exceptions
qdrant_client